fix(QueryEditor): dont render Results every time editor input changes
